* Moved all windows-specific code into pal_win32.h and added pal_linux.h with some stubs * Moved all shader includes to common.h and prefixed all uniform variable name defines, so they can later be moved into the respective headers without causing trouble
