+++ pal_linux.h
@@ -1,0 +1,19 @@
+#ifndef PAL_LINUX_H
+#define PAL_LINUX_H
+
+#include <SDL2/SDL.h>
+
+void ExitProcess(int exit_code) {
+	exit(exit_code);
+}
+
+int main() {
+	//TODO(ca) Implement me
+	return 0;
+}
+
+void flip_buffers() {
+	//TODO(ca) Implement me
+}
+
+#endif
